mzcompress
When you get a moment, can you apply this function to the raw practice data? It just sums the m/z ratios to unit mass. For whatever reason, the raw data reported a higher mass resolution than the instrument was capable of. Maybe it was a poor method, but somehow I doubt it. Attached is also the raw data file as it was output. You can use the mz subdirectory as the input mz in the function. Let me know if you have any questions!